Assassination and vault fix
Assassination is working nicely. The assassination prompt shows up only when the player is crouching and is within distance and disappears when the enemy dies. Their assassination sphere also gets destroyed when they die. May have to change that to get destroyed on health since there may be enemies that get assassinated but are too strong to die from the first hit. Vault is fixed as it no longer vaults over objects that are too high.